backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pydantic[email]==2.8.2
alembic==1.13.2
python-jose==3.3.0
cryptography==42.0.5
